requirements.txt
Jinja2==2.11.2
jsonschema==3.2.0
click==7.1.2
pedalboard==0.7.8
pydub==0.25.1
python-dotenv==1.0.0
pyttsx3==2.90
PyYAML==6.0.1
